+++ tsup.config.ts
@@ -8,7 +8,6 @@ export default defineConfig([
       "host/index": "src/host/index.ts",
       "react/index": "src/react/index.ts",
       "ui/index": "src/ui/index.ts",
-      "ui/fonts": "src/ui/fonts.ts",
       "ui/base": "src/ui/base.ts",
       "vite/index": "src/vite/index.ts",
       "codegen/index": "src/codegen/index.ts",